Move shuffleArray to utilities
