+++ src/utilities/language.ts
@@ -1,4 +1,4 @@
-// Copyright CESSDA ERIC 2017-2023
+// Copyright CESSDA ERIC 2017-2024
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not
 // use this file except in compliance with the License.
@@ -18,6 +18,7 @@
 //   label  : The English label for this language.
 //   index  : The Elasticsearch index containing data for this language.
 // }
+
 export const languages: readonly Language[] = [
   {
     code: "cs",
